seedelf_app/requirements.txt
asgiref==3.8.1
blinker==1.8.2
cached-property==1.5.2
cbor2==5.6.5
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cytoolz==1.0.0
Django==5.1.2
eth-hash==0.7.0
eth-typing==5.0.0
eth-utils==5.0.0
Flask==3.0.3
Flask-Cors==5.0.0
hexbytes==1.2.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
loguru==0.7.2
MarkupSafe==3.0.1
py-ecc==7.0.1
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
toml==0.10.2
toolz==1.0.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
